Parser of Ke-sen Huang's website to obtain website information of papers
This is a parser to fetch webpage information from Ke-Sen Huang's
website and write to their bibtex files. Ke-Sen uses a anti-bot firewall
to block http request, so it is better to manually download the website
and parse them.
I am not sure if it is legal to upload his webpage, so you may have to
do it locally :p
To use the program, download the webpage (e.g.
http://kesen.realtimerendering.com/sig2013.html ) to the same directory
of this file with the bibtex folder, and execute it like this:
python kesen_parser.py http://kesen.realtimerendering.com/sig2013.html
siggraph_2013
where siggraph_2013 is the path point to the folder containing all the
bibtex. The program will automatically find the bibtex for each entry in
Ke-Sen's website and write the website information.

The code is written in a rush and is a quick hack, so it can be super
buggy, be careful!